+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.4.1 (2025-03-21)
+
+### Fix
+
+- Commented multiple `console.log`
+- Added `shouldForwardProp` for styling components avoiding DOM issues.
+- Changed component of Typography to span for fixing the `<p>` decendant of `<p>` issue.
+
 ## 1.4.0 (2025-03-18)
 
 ### Change
@@ -10,7 +18,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Fix
 
-- Fixed wrong notation of width from `chatbotWeight` to `chatbotWidth`.
+- Fixed wrong notation of width from `chatBotWeight` to `chatBotWidth`.
 
 ## 1.3.9 (2025-03-12)